mavlink: move the msg id from uint8_t to uint16_t
With mavlink2 new messages are added with msg IDs greated than 255.
Therefore the msg ID types needed to be raised everywhere.

+++ src/modules/mavlink/mavlink_log_handler.h
@@ -99,7 +99,7 @@ public:
 
 	// Overrides from MavlinkStream
 	const char     *get_name(void) const;
-	uint8_t         get_id(void);
+	uint16_t        get_id(void);
 	unsigned        get_size(void);
 	void            send(const hrt_abstime t);
